backend/requirements.txt
fastapi
uvicorn[standard]
pillow
paddleocr
paddlepaddle
python-multipart
pydantic

gunicorn
requests
python-dotenv
nltk
gevent